requirements.txt
docling==2.56.1
einops==0.8.1
llama-index==0.14.4
llama-index-embeddings-huggingface==0.6.1
llama-index-llms-ollama==0.9.0
nest-asyncio==1.6.0
openai==1.109.1
python-dotenv==1.1.1
qdrant-client==1.15.1
streamlit==1.50.0
transformers==4.57.0
tqdm==4.67.1
